% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoprisca.R
\name{autoprisca}
\alias{autoprisca}
\title{Autoprisca: identical to prisca Algorithm 1 except that L is determined through a heuristic (see paper for details)}
\usage{
autoprisca(
  y,
  a0,
  eps = 0.01,
  grid = FALSE,
  x = NULL,
  n.grid = NULL,
  b0 = NULL,
  ...
)
}
\arguments{
\item{y}{vector of observations}

\item{a0}{hyperparameter on the Gamma distribution ()}

\item{eps}{stopping rule of the ELBO maximization (default is .01)}

\item{grid}{default is FALSE, which corresponds to one observations per time point. If grid=TRUE, it bins the data into a number of intervals specified by n.grid (see below)}

\item{x}{default is NULL. When grid=TRUE, one can inputs at which time instance is y has been collected}

\item{n.grid}{default is NULL. When grid=TRUE, one can inputs how many intervals to define}

\item{b0}{default is equal to a0. It is the rate parameter of the Gaussian}
}
\value{
\code{ELBO}: vector of all ELBO values,
\code{alpha.mat}: It is a L X T matrix. Each row is a posterior of an effect \gamma_l (\alpha_l)
}
\description{
Autoprisca: identical to prisca Algorithm 1 except that L is determined through a heuristic (see paper for details)
}
