% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credible_set.R
\name{credible_set}
\alias{credible_set}
\title{Compute all the credible sets for a given matrix alpha of posterior weights (L x T matrix)}
\usage{
credible_set(
  alpha.mat,
  p = 0.9,
  rm_overlap = TRUE,
  usedelta = FALSE,
  delta = 3
)
}
\arguments{
\item{alpha.mat}{Output of prisca or autoprisca. It is a T x L matrix. Each row is a posterior of an effect \gamma_l (\alpha_l)}

\item{p}{Probability level to build credible sets. Default is 0.9 if not specified}

\item{rm_overlap}{Default is true. Check if there duplicated sets and remove them}

\item{usedelta}{Default is FALSE. If TRUE, it tries to construct set such that the minimum distance between two consecutive times is delta}

\item{delta}{Minimum distance to be used as input if usedelta = TRUE}
}
\value{
\code{cs_list}:A list of credible sets. For each set, we provide: 1) point estimated 2) the set 3) promised coverage of the set
}
\description{
Compute all the credible sets for a given matrix alpha of posterior weights (L x T matrix)
}
