% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prisca_data_sets.R
\name{plot_prisca_data_sets}
\alias{plot_prisca_data_sets}
\title{Plot the data along with the credible sets. It is very much inspired by the corresponding function of the Susie package (Wang et al. 20)}
\usage{
plot_prisca_data_sets(
  out,
  p = 0.9,
  cex = 1.4,
  cex.lab = 1.3,
  cex.axis = 1.3,
  lwd = 3,
  usedelta = FALSE,
  delta = 3,
  option = "data"
)
}
\arguments{
\item{out}{output of prisca or autprisca}

\item{p}{probability level to construct the set used for plotting}

\item{cex}{as in R plot (there is a default)}

\item{cex.lab}{as in R plot (there is a default)}

\item{cex.axis}{as in R plot (there is a default)}

\item{lwd}{as in R plot (there is a default)}

\item{usedelta}{Default is FALSE. If TRUE, it tries to construct set such that the minimum distance between two consecutive times is delta}

\item{delta}{Minimum distance to be used as input if usedelta = TRUE}

\item{option}{It defines where to color the credible sets. Default is "data", meaning that credible sets are coloured directly in the data points. If anything else, colors of the credible sets in the bottom}
}
\value{
\code{plot}
}
\description{
Plot the data along with the credible sets. It is very much inspired by the corresponding function of the Susie package (Wang et al. 20)
}
