% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisca.R
\name{prisca}
\alias{prisca}
\title{Prisca Algorithm 1 implementation}
\usage{
prisca(
  y,
  L,
  a0,
  eps = 0.01,
  limit = 10000,
  grid = FALSE,
  AR = FALSE,
  tf = FALSE,
  x = NULL,
  n.grid = NULL,
  b0 = NULL,
  ...
)
}
\arguments{
\item{y}{vector of observations}

\item{L}{number of single effects (upper bound on changes in variances)}

\item{a0}{hyperparameter on the Gamma distribution ()}

\item{eps}{stopping rule of the ELBO maximization (default is .01)}

\item{limit}{maximum number of iterations for ELBO convengece (default is 10000)}

\item{grid}{default is FALSE, which corresponds to one observations per time point. If grid=TRUE, it bins the data into a number of intervals specified by n.grid (see below)}

\item{AR}{default is FALSE. If AR=TRUE, it fits AR-PRISCA with r=1 (see paper for details)}

\item{tf}{default is FALSE, If AR=TRUE, it fits TF-PRISCA with using the trend filter algorithm (see paper for details)}

\item{x}{default is NULL. When grid=TRUE, one can inputs at which time instance is y has been collected}

\item{n.grid}{default is NULL. When grid=TRUE, one can inputs how many intervals to define}

\item{b0}{default is equal to a0. It is the rate parameter of the Gaussian}
}
\value{
\code{ELBO}: vector of all ELBO values,
\code{alpha.mat}: It is a L x T matrix. Each row is a posterior of an effect \gamma_l (\alpha_l)
\code{phihat}: If AR=TRUE, it outputs the AR coefficient
\code{meanhat}: If tf=TRUE, it outputs the tf estimated mean
}
\description{
Prisca Algorithm 1 implementation
}
