% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priscaXL.R
\name{priscaXL}
\alias{priscaXL}
\title{Prisca Algorithm 1 implementation (XL version, it includes different prior, fractional posterior etc.)}
\usage{
priscaXL(
  y,
  L,
  a0,
  eps = 0.01,
  limit = 10000,
  grid = FALSE,
  AR = FALSE,
  tf = FALSE,
  x = NULL,
  n.grid = NULL,
  b0 = NULL,
  prior = "Gamma",
  alpha.frac = 1,
  ...
)
}
\arguments{
\item{y}{vector of observations}

\item{L}{number of single effects (upper bound on changes in variances)}

\item{a0}{hyperparameter on the Gamma distribution ()}

\item{eps}{stopping rule of the ELBO maximization (default is .01)}

\item{limit}{maximum number of iterations for ELBO convengece (default is 10000)}

\item{grid}{default is FALSE, which corresponds to one observations per time point. If grid=TRUE, it bins the data into a number of intervals specified by n.grid (see below)}

\item{AR}{default is FALSE. If AR=TRUE, it fits AR-PRISCA with r=1 (see paper for details)}

\item{tf}{default is FALSE, If AR=TRUE, it fits TF-PRISCA with using the trend filter algorithm (see paper for details)}

\item{x}{default is NULL. When grid=TRUE, one can inputs at which time instance is y has been collected}

\item{n.grid}{default is NULL. When grid=TRUE, one can inputs how many intervals to define}

\item{b0}{default is equal to a0. It is the rate parameter of the Gaussian}

\item{prior}{default is Gamma prior. Now you can have "Half-Cauchy", if you use the Half Cauchy you need to add an extra parameter gamma. (It can be set to the default value of gamma=25 like in Gelman (2006). A second prior is Gamma-Numeric. It is a gamma prior but the marginal likelihood is computed with Laplace approximation instead of using the closed form (see paper for details)}

\item{alpha.frac}{Parameter of the fractional posterior. Default is 1. It can be set to a number lower than one.}
}
\value{
\code{ELBO}: vector of all ELBO values,
\code{alpha.mat}: It is a L x T matrix. Each row is a posterior of an effect \gamma_l (\alpha_l)
\code{phihat}: If AR=TRUE, it outputs the AR coefficient
\code{meanhat}: If tf=TRUE, it outputs the tf estimated mean
}
\description{
Prisca Algorithm 1 implementation (XL version, it includes different prior, fractional posterior etc.)
}
